+++ api/routes/affiliations/index.js
@@ -1,12 +1,8 @@
 const get = require('./get');
-const post = require('./post');
-const patch = require('./patch');
 
 module.exports = (
   app,
-  { getEndpoint = get, postEndpoint = post, patchEndpoint = patch } = {}
+  { getEndpoint = get } = {}
 ) => {
   getEndpoint(app);
-  postEndpoint(app);
-  patchEndpoint(app);
 };
