+++ api/routes/apds/files/index.js
@@ -1,0 +1,13 @@
+const get = require('./get');
+const post = require('./post');
+
+module.exports = (
+  app,
+  {
+    getEndpoint = get,
+    postEndpoint = post
+  } = {}
+) => {
+  getEndpoint(app);
+  postEndpoint(app)
+};
