+++ api/routes/apds/files/post.js
@@ -15,31 +15,9 @@ module.exports = (
     validateFile = vf,
     createNewFileForAPD = cf,
     deleteFileByID = df,
-    fileBelongsToAPD = fb,
-    getFile = get,
     putFile = put
   } = {}
 ) => {
-  logger.silly('setting up GET /apds/:id/files/:fileID route');
-
-  app.get(
-    '/apds/:id/files/:fileID',
-    can('view-document'),
-    async (req, res, next) => {
-      try {
-        if (await fileBelongsToAPD(req.params.fileID, req.params.id)) {
-          const file = await getFile(req.params.fileID);
-          res.send(file).end();
-        } else {
-          res.status(400).end();
-        }
-      } catch (e) {
-        logger.error({ id: req.id, message: 'error fetching file' });
-        logger.error({ id: req.id, message: e });
-        next(e);
-      }
-    }
-  );
 
   logger.silly('setting up POST /apds/:id/files route');
